✨ feat: 实现h4相关接口
目前实现了 H4 的相关接口，并且单元测试通过，不过目前实现不完全，测试用例覆盖也不全面，后面需要加强。

+++ porting/os/linux/src/os_uart.c
@@ -225,7 +225,7 @@ int os_uart_send(uint8_t *buffer, uint16_t length)
 {
     int len = length;
     while (len > 0) {
-        uint16_t size = write(fd, buffer, len);
+        int size = write(fd, buffer, len);
         if (size <= 0)
             return -1;
         len -= size;
